Phidata/requirements.txt
phidata
python-dotenv
groq
yfinance
duckduckgo-search
googlesearch-python 
pycountry